Save User Evaluation

+++ VacationPackageWebApi/VacationPackageWebApi.Infrastructure.Repositories/Models/RequestOfClient/MainResources/Evaluation/AllAttractionEvaluationPoint.cs
@@ -9,7 +9,7 @@
         }
 
         public Guid Id { get; set; }
-        public short FinalPropertyEvaluation { get; set; }
+        public float FinalPropertyEvaluation { get; set; }
 
         public virtual ICollection<AttractionEvaluation> AttractionEvaluations { get; set; }
         public virtual ICollection<ServiceEvaluation> ServiceEvaluations { get; set; }
